# docs/specifications/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Whiskey Reviews API
  description: |
    Community-driven whiskey review API.

    Reviewers can add, edit, and remove their own whiskey tasting reviews.
    Viewers can browse and read all published reviews.

    - JWT-based authentication with two roles (`reviewer`, `viewer`)
    - Role-based access control (RBAC)
    - Ownership-scoped write operations
    - Pagination on list endpoints
    - Standard error response shapes
  version: 1.0.0
  contact:
    name: Whiskey Reviews API
    email: admin@whiskeyreviews.example.com

servers:
  - url: http://localhost:3000
    description: Local development server

tags:
  - name: Auth
    description: Authentication — register, login, refresh tokens, and logout
  - name: Reviews
    description: Whiskey review management — add, list, view, edit, and remove reviews

paths:
  /v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      description: Create a new user account with either `reviewer` or `viewer` role. Returns access and refresh tokens on success.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'

  /v1/auth/login:
    post:
      tags: [Auth]
      summary: Log in
      description: Authenticate with email and password. Returns access and refresh tokens.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthenticationError'

  /v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: Exchange a refresh token for a new access token and refresh token pair.
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/AuthenticationError'

  /v1/auth/logout:
    post:
      tags: [Auth]
      summary: Log out
      description: Invalidate the current session. Requires a valid bearer token.
      operationId: logoutUser
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Logged out successfully
        '401':
          $ref: '#/components/responses/AuthenticationError'

  /v1/reviews:
    get:
      tags: [Reviews]
      summary: List reviews
      description: Returns a paginated list of all whiskey reviews. Accessible by both `reviewer` and `viewer` roles.
      operationId: listReviews
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Paginated list of reviews
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewList'
        '401':
          $ref: '#/components/responses/AuthenticationError'

    post:
      tags: [Reviews]
      summary: Add a review
      description: Add a new whiskey review. Only `reviewer` role may add reviews. The review is associated with the authenticated reviewer.
      operationId: createReview
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '201':
          description: Review added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

  /v1/reviews/{reviewId}:
    get:
      tags: [Reviews]
      summary: Get a review
      description: Retrieve a single whiskey review by its ID. Accessible by both `reviewer` and `viewer` roles.
      operationId: getReview
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReviewId'
      responses:
        '200':
          description: Review found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      tags: [Reviews]
      summary: Edit a review
      description: Partially edit a whiskey review. Only the `reviewer` who added the review may edit it.
      operationId: updateReview
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReviewId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReviewRequest'
      responses:
        '200':
          description: Review edited successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags: [Reviews]
      summary: Remove a review
      description: Remove a whiskey review. Only the `reviewer` who added the review may remove it.
      operationId: deleteReview
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReviewId'
      responses:
        '204':
          description: Review removed successfully
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    ReviewId:
      name: reviewId
      in: path
      required: true
      description: Unique identifier of the review
      schema:
        type: string
        format: uuid

    Page:
      name: page
      in: query
      required: false
      description: Page number (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1

    PageSize:
      name: pageSize
      in: query
      required: false
      description: Number of results per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  schemas:
    RegisterRequest:
      type: object
      required: [email, password, firstName, lastName, role]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        role:
          type: string
          enum: [reviewer, viewer]

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string

    AuthResponse:
      type: object
      required: [accessToken, refreshToken, expiresIn, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          description: Access token lifetime in seconds
        user:
          $ref: '#/components/schemas/UserSummary'

    UserSummary:
      type: object
      required: [id, email, firstName, lastName, role]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [reviewer, viewer]

    Review:
      type: object
      required: [id, whiskeyName, distillery, region, age, rating, tastingNotes, status, reviewerId, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        whiskeyName:
          type: string
          minLength: 1
        distillery:
          type: string
          minLength: 1
        region:
          type: string
          nullable: true
        age:
          type: integer
          minimum: 1
          nullable: true
          description: Age statement in years; null for no age statement (NAS)
        rating:
          type: integer
          minimum: 1
          maximum: 100
          description: Rating from 1 to 100
        tastingNotes:
          type: string
          nullable: true
        status:
          type: string
          enum: [published, archived]
        reviewerId:
          type: string
          format: uuid
          description: ID of the reviewer who added this review
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ReviewList:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        pagination:
          $ref: '#/components/schemas/Pagination'

    Pagination:
      type: object
      required: [page, pageSize, total]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer

    CreateReviewRequest:
      type: object
      required: [whiskeyName, distillery, rating]
      properties:
        whiskeyName:
          type: string
          minLength: 1
        distillery:
          type: string
          minLength: 1
        region:
          type: string
          nullable: true
        age:
          type: integer
          minimum: 1
          nullable: true
        rating:
          type: integer
          minimum: 1
          maximum: 100
        tastingNotes:
          type: string
          nullable: true

    UpdateReviewRequest:
      type: object
      minProperties: 1
      properties:
        whiskeyName:
          type: string
          minLength: 1
        distillery:
          type: string
          minLength: 1
        region:
          type: string
          nullable: true
        age:
          type: integer
          minimum: 1
          nullable: true
        rating:
          type: integer
          minimum: 1
          maximum: 100
        tastingNotes:
          type: string
          nullable: true
        status:
          type: string
          enum: [published, archived]

    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string

    ValidationError:
      type: object
      required: [code, message, details]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
            required: [field, issue]
            properties:
              field:
                type: string
              issue:
                type: string

  responses:
    ValidationError:
      description: Request body or parameters failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    AuthenticationError:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ForbiddenError:
      description: Authenticated user lacks permission for this operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundError:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ConflictError:
      description: A resource with the same unique identifier already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'